Decouple static initialization and support flushing coverage data to existing ProjectData object.

+++ cobertura/src/test/java/net/sourceforge/cobertura/testutil/Util.java
@@ -94,10 +94,12 @@ public class Util{
 
     public static void removeTestReportFiles(File basedir){
         File[]files = basedir.listFiles();
-        for(File file : files){
-            if(file.getName().startsWith("genericReport") ||
-                    file.getName().equals("cobertura.ser")){
-                file.delete();
+        if(files!=null){
+            for(File file : files){
+                if(file.getName().startsWith("genericReport") ||
+                        file.getName().equals("cobertura.ser")){
+                    file.delete();
+                }
             }
         }
     }
